+++ art/art_param.py
@@ -941,7 +941,17 @@ NON_ASCII_FONTS = [
     "milka",
     "ozana",
     "shanna",
-    "livia"
+    "livia",
+    "lolie",
+    "zakia",
+    "batman",
+    "lord_of_the_ring",
+    "kesia",
+    "akila",
+    "laurine",
+    "bianca",
+    "callian",
+    "awa"
 ]
 
 RANDOM_FILTERED_FONTS = [
@@ -1583,7 +1593,17 @@ FONT_MAP = {"block": [block_dic, True], "banner": [banner_dic, False],  # pragma
             "milka": [milka_dic, False],
             "ozana": [ozana_dic, False],
             "shanna": [shanna_dic, False],
-            "livia": [livia_dic, False]
+            "livia": [livia_dic, False],
+            "lolie": [lolie_dic, False],
+            "zakia": [zakia_dic, False],
+            "batman": [batman_dic, False],
+            "lord_of_the_ring": [lord_of_the_ring_dic, False],
+            "kesia": [kesia_dic, False],
+            "akila": [akila_dic, False],
+            "laurine": [laurine_dic, False],
+            "bianca": [bianca_dic, False],
+            "callian": [callian_dic, False],
+            "awa": [awa_dic, False]
             }
 
 DECORATIONS_MAP = {"angry1": angry1,  # pragma: no cover