+++ CHANGELOG.md
@@ -5,6 +5,21 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- 10 new font
+	1. lolie
+	2. zakia
+	3. batman
+	4. lord_of_the_ring
+	5. kesia
+	6. akila
+	7. laurine
+	8. bianca
+	9. callian
+	10. awa
+### Changed
+- `Python 3.5` support dropped
+- Test system modified
 ## [6.2] - 2024-04-22
 ### Added
 - 5 new font