+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [6.3] - 2024-09-19
 ### Added
 - `data` directory
 - `tests` directory
@@ -2130,7 +2131,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - 1-Line art
 - CLI commands
 
-[Unreleased]: https://github.com/sepandhaghighi/art/compare/v6.2...dev
+[Unreleased]: https://github.com/sepandhaghighi/art/compare/v6.3...dev
+[6.3]: https://github.com/sepandhaghighi/art/compare/v6.2...v6.3
 [6.2]: https://github.com/sepandhaghighi/art/compare/v6.1...v6.2
 [6.1]: https://github.com/sepandhaghighi/art/compare/v6.0...v6.1
 [6.0]: https://github.com/sepandhaghighi/art/compare/v5.9...v6.0